[Removed] file SceneResource : move all methods to Scene, so now scene is a resource, [Added] GameObject + Component to serializer. [Added] After Load method to gameObject + Component

+++ GalaxyEngine/src/Component/MeshComponent.cpp
@@ -137,7 +137,7 @@ namespace GALAXY {
 		{
 			if (!m_materials[i].lock())
 				continue;
-			ImGui::PushID(i);
+			ImGui::PushID((int)i);
 			m_materials[i].lock()->ShowInInspector();
 			ImGui::PopID();
 		}
